+++ backend/benefit4j-starter/src/main/resources/static/assets/benefitPlatformItem-CkJ6zPZz.js
@@ -1,0 +1,1 @@
+import{f as e}from"./index-ByIbUR1w.js";const f=t=>e.get("benefit/api/v1/platform/items",{params:t});export{f as g};